Extracted test_graphing
