+++ corehq/apps/app_manager/tests/test_suite.py
@@ -5,7 +5,6 @@ from django.test import SimpleTestCase
 
 import commcare_translations
 from unittest import mock
-from lxml.etree import tostring
 
 from corehq.apps.app_manager.exceptions import (
     DuplicateInstanceIdError,
@@ -40,7 +39,6 @@ from corehq.apps.app_manager.tests.util import (
     SuiteMixin,
     TestXmlMixin,
     commtrack_enabled,
-    parse_normalize,
     patch_get_xform_resource_overrides,
 )
 from corehq.apps.app_manager.xpath import session_var
@@ -380,33 +378,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin, SuiteMixin):
         ))
 
         self.assertXmlPartialEqual(self.get_xml('update_case_and_subcase'), app.create_suite(), "./entry[1]")
-
-    def test_graphing(self, *args):
-        self._test_generic_suite('app_graphing', 'suite-graphing')
-
-    def test_fixtures_in_graph(self, *args):
-        expected_suite = parse_normalize(self.get_xml('suite-fixture-graphing'), to_string=False)
-        actual_suite = parse_normalize(
-            Application.wrap(self.get_json('app_fixture_graphing')).create_suite(), to_string=False)
-
-        expected_configuration_list = expected_suite.findall('detail/field/template/graph/configuration')
-        actual_configuration_list = actual_suite.findall('detail/field/template/graph/configuration')
-
-        self.assertEqual(len(expected_configuration_list), 1)
-        self.assertEqual(len(actual_configuration_list), 1)
-
-        expected_configuration = expected_configuration_list[0]
-        actual_configuration = actual_configuration_list[0]
-
-        self.assertItemsEqual(
-            [tostring(text_element) for text_element in expected_configuration],
-            [tostring(text_element) for text_element in actual_configuration]
-        )
-
-        expected_suite.find('detail/field/template/graph').remove(expected_configuration)
-        actual_suite.find('detail/field/template/graph').remove(actual_configuration)
-
-        self.assertXmlEqual(tostring(expected_suite), tostring(actual_suite))
 
     def test_printing(self, *args):
         self._test_generic_suite('app_print_detail', 'suite-print-detail')
